refactor(math/fft): move tests to dedicated files and simplify code
- Move inline tests from backend modules to backend/tests.rs
- Move inline tests from gpu/metal modules to fft_tests.rs
- Simplify iterator usage in metal_fft function
- Make internal functions pub(crate) for test access:
  - as_u64_slice_mut
  - compute_root_of_unity
  - goldilocks_square
  - MAX_FFT_ORDER

Test organization:
- backend/tests.rs: CPU/Metal backend differential tests
- gpu/metal/fft_tests.rs: Metal-specific FFT tests
