+++ crypto/math/src/fft/backend/cpu.rs
@@ -50,43 +50,3 @@ impl Fft<GoldilocksField> for CpuFft {
         bowers::batch_ifft::<GoldilocksField>(data, poly_len)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloc::vec::Vec;
-
-    type FE = FieldElement<GoldilocksField>;
-
-    #[test]
-    fn test_cpu_fft_trait_roundtrip() {
-        let backend = CpuFft::new();
-        let original: Vec<FE> = (1..=16).map(|i| FE::from(i as u64)).collect();
-
-        let mut data = original.clone();
-        backend.fft(&mut data).expect("FFT failed");
-
-        // Verify FFT actually transformed the data
-        assert_ne!(data, original, "FFT should transform data");
-
-        backend.ifft(&mut data).expect("IFFT failed");
-        assert_eq!(data, original, "Roundtrip should restore original");
-    }
-
-    #[test]
-    fn test_cpu_fft_trait_batch() {
-        let backend = CpuFft::new();
-        let poly_len = 8;
-        let original: Vec<FE> = (1..=16).map(|i| FE::from(i as u64)).collect();
-
-        let mut data = original.clone();
-        backend
-            .batch_fft(&mut data, poly_len)
-            .expect("Batch FFT failed");
-        backend
-            .batch_ifft(&mut data, poly_len)
-            .expect("Batch IFFT failed");
-
-        assert_eq!(data, original, "Batch roundtrip should restore original");
-    }
-}
