+++ crypto/math/src/fft/backend/metal.rs
@@ -96,9 +96,7 @@ impl Fft<GoldilocksField> for MetalFftBackend {
             return Err(FFTError::InputError(data.len()));
         }
 
-        // For IFFT, we process each polynomial individually since Metal
-        // doesn't have a batch_ifft yet
-        // since the Metal implementation doesn't have batch_ifft yet
+        // Process each polynomial individually since Metal doesn't have batch_ifft
         let n_inv = FieldElement::<GoldilocksField>::from(poly_len as u64)
             .inv()
             .map_err(|_| FFTError::InputError(poly_len))?;
@@ -125,7 +123,7 @@ impl Fft<GoldilocksField> for MetalFftBackend {
 /// - `FieldElement<GoldilocksField>` is a transparent newtype over `u64`
 /// - Both have the same size and alignment
 /// - The conversion is purely a reinterpretation of memory
-fn as_u64_slice_mut(data: &mut [FieldElement<GoldilocksField>]) -> &mut [u64] {
+pub(crate) fn as_u64_slice_mut(data: &mut [FieldElement<GoldilocksField>]) -> &mut [u64] {
     // Compile-time size check
     const _: () = assert!(
         core::mem::size_of::<FieldElement<GoldilocksField>>() == core::mem::size_of::<u64>()
@@ -137,86 +135,4 @@ fn as_u64_slice_mut(data: &mut [FieldElement<GoldilocksField>]) -> &mut [u64] {
     // SAFETY: FieldElement<GoldilocksField> is a #[repr(transparent)] wrapper
     // around u64, so the memory layout is identical
     unsafe { core::slice::from_raw_parts_mut(data.as_mut_ptr() as *mut u64, data.len()) }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use alloc::vec::Vec;
-
-    type FE = FieldElement<GoldilocksField>;
-
-    #[test]
-    fn test_metal_backend_creation() {
-        match MetalFftBackend::try_new() {
-            Ok(_) => println!("Metal backend created successfully"),
-            Err(MetalError::NoDevice) => {
-                println!("No Metal device (expected on non-macOS)");
-            }
-            Err(e) => panic!("Unexpected error: {:?}", e),
-        }
-    }
-
-    #[test]
-    fn test_metal_fft_roundtrip() {
-        let backend = match MetalFftBackend::try_new() {
-            Ok(b) => b,
-            Err(MetalError::NoDevice) => {
-                println!("Skipping: no Metal device");
-                return;
-            }
-            Err(e) => panic!("Unexpected error: {:?}", e),
-        };
-
-        let original: Vec<FE> = (1..=16).map(|i| FE::from(i as u64)).collect();
-        let mut data = original.clone();
-
-        backend.fft(&mut data).expect("FFT failed");
-        assert_ne!(data, original, "FFT should transform data");
-
-        backend.ifft(&mut data).expect("IFFT failed");
-        assert_eq!(data, original, "Roundtrip should restore original");
-    }
-
-    #[test]
-    fn test_metal_batch_fft_roundtrip() {
-        let backend = match MetalFftBackend::try_new() {
-            Ok(b) => b,
-            Err(MetalError::NoDevice) => {
-                println!("Skipping: no Metal device");
-                return;
-            }
-            Err(e) => panic!("Unexpected error: {:?}", e),
-        };
-
-        let poly_len = 8;
-        let original: Vec<FE> = (1..=16).map(|i| FE::from(i as u64)).collect();
-        let mut data = original.clone();
-
-        backend
-            .batch_fft(&mut data, poly_len)
-            .expect("Batch FFT failed");
-        backend
-            .batch_ifft(&mut data, poly_len)
-            .expect("Batch IFFT failed");
-
-        assert_eq!(data, original, "Batch roundtrip should restore original");
-    }
-
-    #[test]
-    fn test_zero_copy_conversion() {
-        // Verify the zero-copy assumption holds
-        let fe = FE::from(12345u64);
-        assert_eq!(core::mem::size_of_val(&fe), core::mem::size_of::<u64>());
-        assert_eq!(*fe.value(), 12345u64);
-
-        // Test the conversion
-        let mut data = alloc::vec![FE::from(1u64), FE::from(2u64), FE::from(3u64)];
-        let raw = as_u64_slice_mut(&mut data);
-        assert_eq!(raw, &[1u64, 2u64, 3u64]);
-
-        // Modify through raw slice
-        raw[1] = 42;
-        assert_eq!(*data[1].value(), 42u64);
-    }
 }
